+++ demos/vinstall.c
@@ -1,9 +1,9 @@
-/* $Id: vinstall.c,v 1.14 2008/10/31 00:11:46 Gregory.Sharp Exp $ */
+/* $Id: vinstall.c,v 1.15 2011/05/15 20:38:16 tom Exp $ */
 
 #include <cdk_test.h>
 
 #ifdef HAVE_XCURSES
-char *XCursesProgramName="vinstall";
+char *XCursesProgramName = "vinstall";
 #endif
 
 /*
@@ -15,37 +15,40 @@ char *XCursesProgramName="vinstall";
 /* Declare global types and prototypes. */
 static char *FPUsage = "-f filename [-s source directory] [-d destination directory] [-t title] [-o Output file] [-q]";
 
-typedef enum {vCanNotOpenSource,
-		vCanNotOpenDest,
-		vOK
-		} ECopyFile;
+typedef enum
+{
+   vCanNotOpenSource,
+   vCanNotOpenDest,
+   vOK
+}
+ECopyFile;
 
 static ECopyFile copyFile (CDKSCREEN *cdkScreen, char *src, char *dest);
 static int verifyDirectory (CDKSCREEN *screen, char *directory);
 
 int main (int argc, char **argv)
 {
-   /* Declare variables. */
-   WINDOW	*cursesWin	= 0;
-   CDKSCREEN	*cdkScreen	= 0;
-   CDKSWINDOW	*installOutput	= 0;
-   CDKENTRY	*sourceEntry	= 0;
-   CDKENTRY	*destEntry	= 0;
-   CDKLABEL	*titleWin	= 0;
-   CDKHISTOGRAM *progressBar	= 0;
-   char		*sourcePath	= 0;
-   char		*destPath	= 0;
-   char		*sourceDir	= 0;
-   char		*destDir	= 0;
-   char		*filename	= 0;
-   char		*title		= 0;
-   char		*output		= 0;
-   int		quiet		= FALSE;
-   int		errors		= 0;
-   int		sWindowHeight	= 0;
-   char		*titleMessage[10], **fileList = 0, *mesg[20];
-   char		oldPath[512], newPath[512], temp[2000];
-   char		**files;
+   /* *INDENT-EQLS* */
+   WINDOW *cursesWin            = 0;
+   CDKSCREEN *cdkScreen         = 0;
+   CDKSWINDOW *installOutput    = 0;
+   CDKENTRY *sourceEntry        = 0;
+   CDKENTRY *destEntry          = 0;
+   CDKLABEL *titleWin           = 0;
+   CDKHISTOGRAM *progressBar    = 0;
+   char *sourcePath             = 0;
+   char *destPath               = 0;
+   char *sourceDir              = 0;
+   char *destDir                = 0;
+   char *filename               = 0;
+   char *title                  = 0;
+   char *output                 = 0;
+   int quiet                    = FALSE;
+   int errors                   = 0;
+   int sWindowHeight            = 0;
+   char *titleMessage[10], **fileList = 0, *mesg[20];
+   char oldPath[512], newPath[512], temp[2000];
+   char **files;
    int count, chunks, ret, x;
 
    /* Parse up the command line. */
@@ -58,29 +61,29 @@ int main (int argc, char **argv)
       }
       switch (ret)
       {
-	 case 's' :
-	      sourcePath = strdup (optarg);
-	      break;
+      case 's':
+	 sourcePath = strdup (optarg);
+	 break;
 
-	 case 'd' :
-	      destPath = strdup (optarg);
-	      break;
+      case 'd':
+	 destPath = strdup (optarg);
+	 break;
 
-	 case 'f' :
-	      filename = strdup (optarg);
-	      break;
+      case 'f':
+	 filename = strdup (optarg);
+	 break;
 
-	 case 't' :
-	      title = strdup (optarg);
-	      break;
+      case 't':
+	 title = strdup (optarg);
+	 break;
 
-	 case 'o' :
-	      output = strdup (optarg);
-	      break;
+      case 'o':
+	 output = strdup (optarg);
+	 break;
 
-	 case 'q' :
-	      quiet = TRUE;
-	      break;
+      case 'q':
+	 quiet = TRUE;
+	 break;
       }
    }
 
@@ -99,22 +102,22 @@ int main (int argc, char **argv)
       ExitProgram (EXIT_FAILURE);
    }
 
-  /*
-   * Cycle through what was given to us and save it.
-   */
-   for (x=0; x < count; x++)
+   /*
+    * Cycle through what was given to us and save it.
+    */
+   for (x = 0; x < count; x++)
    {
       /* Strip white space from the line. */
       stripWhiteSpace (vBOTH, fileList[x]);
    }
 
-   /* Set up CDK. */ 
-   cursesWin = initscr();
+   /* Set up CDK. */
+   cursesWin = initscr ();
    cdkScreen = initCDKScreen (cursesWin);
 
    /* Start color. */
-   initCDKColor();
-   
+   initCDKColor ();
+
    /* Create the title label. */
    titleMessage[0] = "<C></32/B><#HL(30)>";
    if (title == 0)
@@ -127,23 +130,23 @@ int main (int argc, char **argv)
    }
    titleMessage[1] = copyChar (temp);
    titleMessage[2] = "<C></32/B><#HL(30)>";
-   titleWin = newCDKLabel (cdkScreen, CENTER, TOP, 
-				titleMessage, 3, FALSE, FALSE);
+   titleWin = newCDKLabel (cdkScreen, CENTER, TOP,
+			   titleMessage, 3, FALSE, FALSE);
    freeChar (titleMessage[1]);
 
    /* Allow them to change the install directory. */
    if (sourcePath == 0)
    {
       sourceEntry = newCDKEntry (cdkScreen, CENTER, 8,
-					0, "Source Directory        :",
-					A_NORMAL, '.', vMIXED, 
-					40, 0, 256, TRUE, FALSE);
+				 0, "Source Directory        :",
+				 A_NORMAL, '.', vMIXED,
+				 40, 0, 256, TRUE, FALSE);
    }
    if (destPath == 0)
    {
       destEntry = newCDKEntry (cdkScreen, CENTER, 11,
-				0, "Destination Directory:", A_NORMAL, 
-				'.', vMIXED, 40, 0, 256, TRUE, FALSE);
+			       0, "Destination Directory:", A_NORMAL,
+			       '.', vMIXED, 40, 0, 256, TRUE, FALSE);
    }
 
    /* Get the source install path. */
@@ -178,9 +181,9 @@ int main (int argc, char **argv)
       destroyCDKEntry (destEntry);
    }
 
-  /*
-   * Verify that the source directory is valid.
-   */
+   /*
+    * Verify that the source directory is valid.
+    */
    if (verifyDirectory (cdkScreen, sourceDir) != 0)
    {
       /* Clean up and leave. */
@@ -188,13 +191,13 @@ int main (int argc, char **argv)
       freeChar (sourceDir);
       destroyCDKLabel (titleWin);
       destroyCDKScreen (cdkScreen);
-      endCDK();
+      endCDK ();
       ExitProgram (EXIT_FAILURE);
    }
 
-  /*
-   * Verify that the source directory is valid.
-   */
+   /*
+    * Verify that the source directory is valid.
+    */
    if (verifyDirectory (cdkScreen, destDir) != 0)
    {
       /* Clean up and leave. */
@@ -202,25 +205,25 @@ int main (int argc, char **argv)
       freeChar (sourceDir);
       destroyCDKLabel (titleWin);
       destroyCDKScreen (cdkScreen);
-      endCDK();
+      endCDK ();
       ExitProgram (EXIT_FAILURE);
    }
 
    /* Create the histogram. */
    progressBar = newCDKHistogram (cdkScreen, CENTER, 5,
-					3, 0, HORIZONTAL,
-					"<C></56/B>Install Progress",
-					TRUE, FALSE);
+				  3, 0, HORIZONTAL,
+				  "<C></56/B>Install Progress",
+				  TRUE, FALSE);
 
-   /* Set the top left/right characters of the histogram.*/
+   /* Set the top left/right characters of the histogram. */
    setCDKHistogramLLChar (progressBar, ACS_LTEE);
    setCDKHistogramLRChar (progressBar, ACS_RTEE);
 
    /* Set the initial value of the histogram. */
    setCDKHistogram (progressBar, vPERCENT, TOP, A_BOLD,
-			1, count, 1,
-			COLOR_PAIR (24) | A_REVERSE | ' ',
-			TRUE);
+		    1, count, 1,
+		    COLOR_PAIR (24) | A_REVERSE | ' ',
+		    TRUE);
 
    /* Determine the height of the scrolling window. */
    if (LINES >= 16)
@@ -231,14 +234,14 @@ int main (int argc, char **argv)
    {
       sWindowHeight = 3;
    }
-   
+
    /* Create the scrolling window. */
    installOutput = newCDKSwindow (cdkScreen, CENTER, BOTTOM,
-					sWindowHeight, 0,
-					"<C></56/B>Install Results",
-					2000, TRUE, FALSE);
+				  sWindowHeight, 0,
+				  "<C></56/B>Install Results",
+				  2000, TRUE, FALSE);
 
-   /* Set the top left/right characters of the scrolling window.*/
+   /* Set the top left/right characters of the scrolling window. */
    setCDKSwindowULChar (installOutput, ACS_LTEE);
    setCDKSwindowURChar (installOutput, ACS_RTEE);
 
@@ -246,15 +249,15 @@ int main (int argc, char **argv)
    drawCDKScreen (cdkScreen);
 
    /* Start copying the files. */
-   for (x=0; x < count; x++)
+   for (x = 0; x < count; x++)
    {
-     /*
-      * If the 'file' list file has 2 columns, the first is
-      * the source filename, the second being the destination
-      * filename.
-      */
+      /*
+       * If the 'file' list file has 2 columns, the first is
+       * the source filename, the second being the destination
+       * filename.
+       */
       files = CDKsplitString (fileList[x], ' ');
-      chunks = CDKcountStrings (files);
+      chunks = (int)CDKcountStrings (files);
       if (chunks == 2)
       {
 	 /* Create the correct paths. */
@@ -267,18 +270,20 @@ int main (int argc, char **argv)
 	 sprintf (oldPath, "%s/%s", sourceDir, fileList[x]);
 	 sprintf (newPath, "%s/%s", destDir, fileList[x]);
       }
-      CDKfreeStrings(files);
+      CDKfreeStrings (files);
 
       /* Copy the file from the source to the destination. */
       ret = copyFile (cdkScreen, oldPath, newPath);
       if (ret == vCanNotOpenSource)
       {
-	 sprintf (temp, "</16>Error: Can not open source file \"%.256s\"<!16>", oldPath);
+	 sprintf (temp,
+		  "</16>Error: Can not open source file \"%.256s\"<!16>", oldPath);
 	 errors++;
       }
       else if (ret == vCanNotOpenDest)
       {
-	 sprintf (temp, "</16>Error: Can not open destination file \"%.256s\"<!16>", newPath);
+	 sprintf (temp,
+		  "</16>Error: Can not open destination file \"%.256s\"<!16>", newPath);
 	 errors++;
       }
       else
@@ -288,26 +293,27 @@ int main (int argc, char **argv)
 
       /* Add the message to the scrolling window. */
       addCDKSwindow (installOutput, temp, BOTTOM);
-      drawCDKSwindow (installOutput, ObjOf(installOutput)->box);
+      drawCDKSwindow (installOutput, ObjOf (installOutput)->box);
 
       /* Update the histogram. */
       setCDKHistogram (progressBar, vPERCENT, TOP, A_BOLD,
-			1, count, x+1, 
-			COLOR_PAIR (24) | A_REVERSE | ' ',
-			TRUE);
-      
+		       1, count, x + 1,
+		       COLOR_PAIR (24) | A_REVERSE | ' ',
+		       TRUE);
+
       /* Update the screen. */
       drawCDKHistogram (progressBar, TRUE);
    }
 
-  /*
-   * If there were errors, inform the user and allow them to look at the
-   * errors in the scrolling window.
-   */
+   /*
+    * If there were errors, inform the user and allow them to look at the
+    * errors in the scrolling window.
+    */
    if (errors != 0)
    {
       /* Create the information for the dialog box. */
-      char *buttons[] = {"Look At Errors Now", "Save Output To A File", "Ignore Errors"};
+      char *buttons[] =
+      {"Look At Errors Now", "Save Output To A File", "Ignore Errors"};
       mesg[0] = "<C>There were errors in the installation.";
       mesg[1] = "<C>If you want, you may scroll through the";
       mesg[2] = "<C>messages of the scrolling window to see";
@@ -330,20 +336,21 @@ int main (int argc, char **argv)
    }
    else
    {
-     /*
-      * If they specified the name of an output file, then save the
-      * results of the installation to that file.
-      */
+      /*
+       * If they specified the name of an output file, then save the
+       * results of the installation to that file.
+       */
       if (output != 0)
       {
-	  dumpCDKSwindow (installOutput, output);
+	 dumpCDKSwindow (installOutput, output);
       }
       else
       {
 	 /* Ask them if they want to save the output of the scrolling window. */
 	 if (quiet == FALSE)
 	 {
-	    char *buttons[] = {"No", "Yes"};
+	    char *buttons[] =
+	    {"No", "Yes"};
 	    mesg[0] = "<C>Do you want to save the output of the";
 	    mesg[1] = "<C>scrolling window to a file?";
 
@@ -354,13 +361,13 @@ int main (int argc, char **argv)
 	 }
       }
    }
-   
+
    /* Clean up. */
    destroyCDKLabel (titleWin);
    destroyCDKHistogram (progressBar);
    destroyCDKSwindow (installOutput);
    destroyCDKScreen (cdkScreen);
-   endCDK();
+   endCDK ();
    ExitProgram (EXIT_SUCCESS);
 }
 
@@ -380,10 +387,10 @@ static ECopyFile copyFile (CDKSCREEN *cdkScreen GCC_UNUSED, char *src, char *des
    }
    fclose (fd);
 
-  /*
-   * Remove the destination file first, just in case it already exists.
-   * This allows us to check if we can write to the desintation file.
-   */
+   /*
+    * Remove the destination file first, just in case it already exists.
+    * This allows us to check if we can write to the desintation file.
+    */
    sprintf (command, "rm -f %s", dest);
    system (command);
 
@@ -394,12 +401,12 @@ static ECopyFile copyFile (CDKSCREEN *cdkScreen GCC_UNUSED, char *src, char *des
    }
    fclose (fd);
 
-  /*
-   * Copy the file. There has to be a better way to do this. I
-   * tried rename and link but they both have the same limitation
-   * as the 'mv' command that you can not move across partitions.
-   * Quite limiting in an install binary.
-   */
+   /*
+    * Copy the file. There has to be a better way to do this. I
+    * tried rename and link but they both have the same limitation
+    * as the 'mv' command that you can not move across partitions.
+    * Quite limiting in an install binary.
+    */
    sprintf (command, "rm -f %s; cp %s %s; chmod 444 %s", dest, src, dest, dest);
    system (command);
    return vOK;
@@ -412,10 +419,11 @@ static ECopyFile copyFile (CDKSCREEN *cdkScreen GCC_UNUSED, char *src, char *des
  */
 static int verifyDirectory (CDKSCREEN *cdkScreen, char *directory)
 {
-   char *buttons[]	= {"Yes", "No"};
-   int status		= 0;
+   char *buttons[] =
+   {"Yes", "No"};
+   int status = 0;
 #if !defined (__MINGW32__)
-   mode_t dirMode	= S_IRWXU | S_IRGRP | S_IXGRP | S_IROTH;
+   mode_t dirMode = S_IRWXU | S_IRGRP | S_IXGRP | S_IROTH;
 #endif
    struct stat fileStat;
    char *mesg[10];
@@ -470,14 +478,14 @@ static int verifyDirectory (CDKSCREEN *cdkScreen, char *directory)
 	 {
 	    /* Create the message. */
 	    error[0] = "<C>Installation aborted.";
-	    
+
 	    /* Pop up the error message. */
 	    popupLabel (cdkScreen, error, 1);
 
 	    /* Set the exit status. */
 	    status = -1;
 	 }
- 
+
 	 /* Clean up. */
 	 freeChar (mesg[1]);
       }
